+++ cli/BatchReader.h
@@ -1,8 +1,9 @@
 #pragma once
 
 #include <fstream>
-#include <string>
-#include <vector>
+#include <mutex>
+
+#include "Batch.h"
 
 class BatchReader
 {
@@ -10,10 +11,18 @@ public:
   BatchReader(const std::string& file, size_t batch_size);
   BatchReader(std::istream& in, size_t batch_size);
 
-  std::vector<std::string> read_next();
+  Batch read_next();
+
+  size_t read_lines() const
+  {
+    return _read_lines;
+  }
 
 private:
   std::ifstream _file;
   std::istream& _in;
   size_t _batch_size;
+  size_t _batch_id;
+  size_t _read_lines;
+  std::mutex _reader_mutex;
 };