+++ cli/BatchWriter.cc
@@ -3,16 +3,38 @@
 BatchWriter::BatchWriter(const std::string& file)
   : _file(file.c_str())
   , _out(_file)
+  , _last_batch_id(0)
 {
 }
 
 BatchWriter::BatchWriter(std::ostream& out)
   : _out(out)
+  , _last_batch_id(0)
 {
 }
 
-void BatchWriter::write(const std::vector<std::string>& batch)
+void BatchWriter::write(const Batch& batch)
 {
-  for (const auto& sent: batch)
-    _out << sent << std::endl;
+  std::lock_guard<std::mutex> lock(_writer_mutex);
+  size_t batch_id = batch.get_id();
+  if (batch_id == _last_batch_id + 1)
+  {
+    std::vector<std::string> v = batch.get_data();
+
+    while (batch_id == _last_batch_id + 1)
+    {
+      for (const auto& sent: v)
+        _out << sent << std::endl;
+      _last_batch_id = batch_id;
+      auto it = _pending_batches.find(_last_batch_id + 1);
+      if (it != _pending_batches.end())
+      {
+        v = it->second;
+        batch_id = _last_batch_id + 1;
+        _pending_batches.erase(it);
+      }
+    }
+  }
+  else
+    _pending_batches[batch_id] = batch.get_data();
 }