+++ cli/BatchWriter.h
@@ -1,8 +1,10 @@
 #pragma once
 
 #include <fstream>
-#include <string>
-#include <vector>
+#include <mutex>
+#include <map>
+
+#include "Batch.h"
 
 class BatchWriter
 {
@@ -10,9 +12,12 @@ public:
   BatchWriter(const std::string& file);
   BatchWriter(std::ostream& out);
 
-  void write(const std::vector<std::string>& batch);
+  void write(const Batch& batch);
 
 private:
   std::ofstream _file;
   std::ostream& _out;
+  std::map<size_t, std::vector<std::string>> _pending_batches;
+  size_t _last_batch_id;
+  std::mutex _writer_mutex;
 };